Add Bootstrap

+++ src/main/java/org/project/notablog/repos/MessageRepo.java
@@ -4,7 +4,9 @@ import org.project.notablog.domains.Message;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface MessageRepo extends CrudRepository<Message, Long> {
+public interface MessageRepo extends CrudRepository<Message, Integer> {
     List<Message> findByTag(String tag);
+
 }
